resolve some errors and split code
