+++ README.md
@@ -3,6 +3,11 @@
 piscon用のポータルサイト
 
 memo
+
+中途半端に分割してるからきれいにしたい
+Conohaのほうも書き換えないと分離の意味がない…
+(まあこれ以降Conoha使わん気がやや)
+
 ENV
 ```
 ACCESS_ID
@@ -14,4 +19,4 @@ MARIADB_DATABASE
 IF AWS
 
 AWS_SUBNET_ID
-
+```
